CalculateLaneDistance had multiplier 1/8.0 instead of 1/80. This was causing the engine to produce awkward traffic. Also changed the return value from double to float
git-svn-id: http://project-liberty.googlecode.com/svn/trunk@40 562435f6-89a1-033f-70e2-f487d5484f5a

+++ SOLCore/GameVC/CPathFind.h
@@ -117,7 +117,7 @@ public:
     _pad(__fxpad00, 1); // 11-12
 
     CDetachedNode();
-    double CalculateLaneDistance();
+    float CalculateLaneDistance();
 };
 
 class CPathFind{
